requirements.txt
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
excel-base==1.0.4
isoweek==1.3.3
python-dateutil==2.9.0.post0
six==1.16.0
TimeConvert==3.0.13
tzlocal==5.2
xlwt==1.3.0
